+++ src/database/entities/members.rs
@@ -7,8 +7,10 @@ use sea_orm::entity::prelude::*;
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i64,
-    pub guild: u64,
-    pub user: u64,
+    #[sea_orm(column_type = "Decimal(Some((20, 0)))")]
+    pub guild: Decimal,
+    #[sea_orm(column_type = "Decimal(Some((20, 0)))")]
+    pub user: Decimal,
     pub points: i64,
 }
